+++ src/components/DatePicker/datePickerChangelogData.ts
@@ -13,6 +13,15 @@ export const changelogData: ChangelogData[] = [
     date: "Prerelease",
     version: "Prerelease",
     type: "Update",
+    affects: ["Styles"],
+    notes: [
+      "Fixes bug where month and year calendars were rendering vertically rather than horizontally.",
+    ],
+  },
+  {
+    date: "Prerelease",
+    version: "Prerelease",
+    type: "Update",
     affects: ["Functionality"],
     notes: [
       "Updates the `initialDate` and `initialDateTo` props to accept an empty string and adds optional `placeholder` and `placeholderTo` props.",